// admin_pannel/src/@types/interface/retailer.interface.ts
export interface IRetailers {
    retailer_name: string;
	contact_name: string;
	contact_phone: string;
	contact_email: string;
	trade_license_number: string;
	nid_number: string;
    logo: string;
	sign_board_photo: string;
	retailer_owner_photo: string;
	trade_license_photo: string;
	nid_photo: string;
}